simulation/requirements.txt
pymodbus>=2.5.3
numpy>=1.20.3
pandas>=1.3.4
PyYAML>=6.0
python-dotenv>=0.19.2
pyserial>=3.5 